move CommitRef enum to its own module
