+++ fluree-db-api/src/lib.rs
@@ -36,6 +36,7 @@
 pub mod admin;
 pub mod block_fetch;
 pub mod bm25_worker;
+pub mod commit_ref;
 pub mod commit_transfer;
 pub mod config_resolver;
 #[cfg(feature = "credential")]
